Using Fireblocks Tokenization endpoints
