+++ fireblocks_sdk/__init__.py
@@ -1,6 +1,5 @@
 from fireblocks_sdk.sdk import FireblocksSDK
 from fireblocks_sdk.ncw_sdk import FireblocksNCW
-from fireblocks_sdk.tokenization_sdk import FbTokenizationSDK
 from fireblocks_sdk.sdk_token_provider import SdkTokenProvider
 from fireblocks_sdk.api_types import *
 from fireblocks_sdk.tokenization_api_types import *
